Ajustado o campo float

+++ src/forms/user-profile-form/manage-restaurant-form/ManageRestaurantForm.tsx
@@ -70,13 +70,13 @@ const ManageRestaurantForm = ({ onSave, restaurant, isLoading }: Props) => {
       return;
     }
 
-    const deliveryPriceFormatted = parseInt(
+    const deliveryPriceFormatted = parseFloat(
       (restaurant.deliveryPrice / 100).toFixed(2)
     );
 
     const menuItemsFormatted = restaurant.menuItems.map((item) => ({
       ...item,
-      price: parseInt((item.price / 100).toFixed(2)),
+      price: parseFloat((item.price / 100).toFixed(2)),
     }));
 
     const updatedRestaurant = {
